remove old blueprint
