+++ examples/chalicelib/blueprints/__init__.py
@@ -1,13 +1,9 @@
 __all__ = ['AuthedRestApiBlueprint']
 
-from agave.blueprints import RestApiBlueprint, RestApiBlueprintV2
+from agave.blueprints import RestApiBlueprint
 
 from .authed import AuthedBlueprint
 
 
 class AuthedRestApiBlueprint(AuthedBlueprint, RestApiBlueprint):
-    ...
-
-
-class AuthedRestApiBlueprintV2(AuthedBlueprint, RestApiBlueprintV2):
     ...
