+++ examples/chalicelib/resources/__init__.py
@@ -1,14 +1,9 @@
 __all__ = [
     'app',
-    'app_v2',
     'Account',
-    'AccountV2',
     'Transaction',
-    'TransactionV2',
 ]
 
 from .accounts import Account
-from .accounts_v2 import AccountV2
-from .base import app, app_v2
+from .base import app
 from .transactions import Transaction
-from .transactions_v2 import TransactionV2
