+++ examples/chalicelib/resources/transactions.py
@@ -1,4 +1,5 @@
 from agave.filters import generic_query
+from agave.repositories import MongoRepository
 
 from ..models.transactions import Transaction as TransactionModel
 from ..validators import TransactionQuery
@@ -7,6 +8,6 @@ from .base import app
 
 @app.resource('/transactions')
 class Transaction:
-    model = TransactionModel
+    repository = MongoRepository(TransactionModel, generic_query)
     query_validator = TransactionQuery
     get_query_filter = generic_query
